show_homepage/requirements.txt
google-cloud-storage==1.10.0
bs4